requirements.txt
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.41.0
django-tof==0.3.2
gunicorn==20.0.4
oauth==1.0.1
oauthlib==3.1.0
Pillow==6.2.2
psycopg2==2.8.5
whitenoise==5.0.1
